requirements.txt
Django==1.9.5
django-redis-sessions==0.5.0
django-websocket-redis==0.4.6
djangorestframework==3.3.3
drf-nested-routers==0.11.1
gevent==1.1.1
greenlet==0.4.9
Markdown==2.6.6
redis==2.10.5
six==1.10.0
wsaccel==0.6.2
